requirements.txt
Flask
requests
python-dotenv
spotipy
sentence-transformers
transformers
torch
pandas
numpy
scikit-learn
tqdm
langdetect
lyricsgenius
werkzeug